ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
scikit-learn
xgboost
joblib
